+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 - Missing conversion methods, and standardized the existing ones (see the
   library docs for a list of conversion method names)
+- Instant now has a `parse()` constructor for ISO 8601 formatted strings
 
 ### Changed